+++ content-page.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * The template used for displaying page content in page.php
+ *
+ * @package _mbbasetheme
+ */
+?>
+
+<article id="post-<?php the_ID(); ?>" <?php post_class(); ?>>
+	<header class="entry-header">
+		<h1 class="entry-title"><?php the_title(); ?></h1>
+	</header><!-- .entry-header -->
+
+	<div class="entry-content">
+		<?php the_content(); ?>
+		<?php
+			wp_link_pages( array(
+				'before' => '<div class="page-links">' . __( 'Pages:', '_mbbasetheme' ),
+				'after'  => '</div>',
+			) );
+		?>
+	</div><!-- .entry-content -->
+	<?php edit_post_link( __( 'Edit', '_mbbasetheme' ), '<footer class="entry-footer"><span class="edit-link">', '</span></footer>' ); ?>
+</article><!-- #post-## -->