+++ footer.php
@@ -1,11 +1,15 @@
-		<footer>
-			<?php dynamic_sidebar( 'Footer' ); ?>
+	</div><!-- #content -->
+
+	<footer id="colophon" class="site-footer" role="contentinfo">
+		<?php dynamic_sidebar( 'Footer' ); ?>
+		<div class="site-info">
 			<p>&copy; <?php echo date( "Y" ); echo " "; bloginfo( 'name' ); ?></p>
-		</footer>
+		</div><!-- .site-info -->
+	</footer><!-- #colophon -->
 
-	</div> <!-- /#container -->
+</div><!-- #page -->
 
-  <?php wp_footer(); ?>
+<?php wp_footer(); ?>
 
 </body>
 </html>