+++ functions.php
@@ -1,13 +1,53 @@
 <?php
+/**
+ * _mbbasetheme functions and definitions
+ *
+ * @package _mbbasetheme
+ */
 
 /****************************************
 Theme Setup
 *****************************************/
 
+/**
+ * Theme initialization
+ */
 require_once( get_template_directory() . '/lib/init.php' );
-require_once( get_template_directory() . '/lib/theme-helpers.php' );
+
+/**
+ * Custom theme functions definited in /lib/init.php
+ */
 require_once( get_template_directory() . '/lib/theme-functions.php' );
-require_once( get_template_directory() . '/lib/theme-comments.php' );
+
+/**
+ * Helper functions for use in other areas of the theme
+ */
+require_once( get_template_directory() . '/lib/theme-helpers.php' );
+
+/**
+ * Implement the Custom Header feature.
+ */
+//require get_template_directory() . 'lib/inc/custom-header.php';
+
+/**
+ * Custom template tags for this theme.
+ */
+require get_template_directory() . 'lib/inc/template-tags.php';
+
+/**
+ * Custom functions that act independently of the theme templates.
+ */
+require get_template_directory() . 'lib/inc/extras.php';
+
+/**
+ * Customizer additions.
+ */
+require get_template_directory() . 'lib/inc/customizer.php';
+
+/**
+ * Load Jetpack compatibility file.
+ */
+require get_template_directory() . 'lib/inc/jetpack.php';
 
 
 /****************************************
@@ -27,6 +67,7 @@ Misc Theme Functions
 /**
  * Define custom post type capabilities for use with Members
  */
+add_action( 'admin_init', 'mb_add_post_type_caps' );
 function mb_add_post_type_caps() {
 	// mb_add_capabilities( 'portfolio' );
 }